Rename the xref-backend-extra-* methods to use the term definition
Because they are not "extra", and to a large extent will contain
"kinds" already included in the 'xref-find-definitions' output.

* lisp/progmodes/eglot.el (eglot--xref-kinds-alist): Remove the entry
for 'references' because of differing semantics.  Simplify the names
of the keys as well (remove the 'eglot--xref-' prefix) because these
are shown to the user.  They won't be sent to any other backend, so
private names seem unnecessary.  Use strings, to ensure specific type.

* lisp/progmodes/eglot.el (xref-backend-definitions-by-kind):
Skip 'intern'.

* lisp/progmodes/xref.el (xref-backend-definition-kinds):
Rename from 'xref-backend-extra-kinds'.  Update all references.
(xref-xref-backend-definitions-by-kind): Rename from
'xref-backend-extra-defs'.  Update all references.