(vc-print-log-setup-buttons): Start "previous" history with specified revision
* lisp/vc/vc-git.el (vc-git-file-name-changes-switches): Remove
the comment above the option.  Seems unnecessary now.

* lisp/vc/vc.el (vc-print-log-setup-buttons): Start the "previous"
change history buffer with the specified revision, rather than
have the sentinel jump to it.  Apparently in some cases the
history of the old name can't be found.  In others, the log just
shows faster.  But note the caveat described in the second new
comment (bug#55871).