In a multi-display setup, the main menu does not move from one display to another when the Emacs window is moved between displays.  This improvement adds a notification listener that notifies the mainMenu that the NSWindowDidChangeScreenNotification has occured.
In other Apps this callback is registered automatically.  In NSMenu, the
applicationDidFinishLaunch method registers this for the mainMenu.
However, the way emacs construcgts GNUstep menus in nsmenu.m, constructs
menus in a different manner.

Note: this proposed modification handles one emacs window moving between
different displays, but does not handle the case where there are multiple
emacs windows.  In that case, the menu should appear on the display on
which the emacs frame is shown.